Split elmu-server into dedicated controller modules
* Also fix wrong table header in user edit page

Resolves #34
